+++ src/game/rendering_graph_node.c
@@ -325,9 +325,9 @@ static void geo_process_perspective(struct GraphNodePerspective *node) {
         gSpeedrunTimer++;
     }
     if (gSpeedrunMode) {
-        u64 timerMinutes = gSpeedrunTimer / (30 * 60);
-        u64 timerSeconds = (gSpeedrunTimer - (timerMinutes * 1800)) / 30;
-        u64 fracSecs = ((gSpeedrunTimer - (timerMinutes * 1800) - (timerSeconds * 30)) & 0xFFFF) / 3;
+        u32 timerMinutes = gSpeedrunTimer / (30 * 60);
+        u32 timerSeconds = (gSpeedrunTimer - (timerMinutes * 1800)) / 30;
+        u32 fracSecs = ((gSpeedrunTimer - (timerMinutes * 1800) - (timerSeconds * 30)) & 0xFFFF) / 3;
         if (timerMinutes < 100) {
             print_text_fmt_int(GFX_DIMENSIONS_RECT_FROM_RIGHT_EDGE(102), 210, "%02d", timerMinutes);
         } else {
